move np plugins and mpl plugins to a separate dicts
initial step toward separation from np and mpl
