+++ src/nxscli/mpl/plot_mpl.py
