+++ src/nxscli/mpl/plugins/capture.py
@@ -8,7 +8,7 @@ from nxscli.iplugin import IPluginPlotStatic
 from nxscli.logger import logger
 from nxscli.main.environment import Environment, pass_environment
 from nxscli.main.types import Samples, plot_options
-from nxscli.plot_mpl import MplManager, PluginPlotMpl
+from nxscli.mpl.plot_mpl import MplManager, PluginPlotMpl
 from nxscli.pluginthr import PluginThread
 
 if TYPE_CHECKING:
