+++ src/nxscli/np/plugins/npmem.py
