+++ src/nxscli/np/plugins/npsave.py
